backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
ib-insync==0.9.86
python-dotenv==1.0.1
httpx==0.27.2
alpaca-trade-api==3.2.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
cryptography==43.0.3
python-multipart==0.0.21
google-genai>=1.0.0
